# CategoryTreeSorting.init jquery UI performance